[13.0][FIX] mis_builder: always ignore cancel entries.
When you select all entries in "Target Moves", the expected is
that the report considers both posted and draft entries.

In v13 and due to the merge of invoice in `account.move` model,
a new state `cancel` has been introduced. This forces to always
filter entries in this state to keep the behavior of previous
versions (and expected).